# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Share the moments you love with people you love!
  version: 2.0.2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    uid:
      name: uid
      description: User identifier given by server at the moment of the sign up
      in: path
      schema:
        $ref: "#/components/schemas/user/properties/id"
      required: true
    photo_id:
      name: photo_id
      in: path
      description: Photo id
      schema:
        type: integer
        description: Photo id
        example: 12
      required: true
    comment_id:
      name: comment_id
      description: Comment id
      in: path
      schema:
        type: integer
        description: Comment id
        example: 12
      required: true
  schemas:
    photo:
      type: object
      description: Photo Object
      properties:
        photo_data:
          type: string
          format: binary
          description: photo binary data
          minLength: 0
          maxLength: 100
        username_owner:
          $ref: "#/components/schemas/user/properties/name"
...
    comment:
      required:
        - text
        - username_owner
      type: object
      description: Comment object
      properties:
        text:
          type: string
          description: 'Text of a comment'
          pattern: '^.{3,150}$'
          example: "This is a comment!"
          minLength: 1
          maxLength: 150
        username_owner:
          $ref: "#/components/schemas/user/properties/name"
  responses:
    ServerDown:
      description: "Server is down"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Server Down"
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Bad request"
...
    DeleteSuccess:
      description: "Resource deleted successfully"
      content:
        application/json:
          schema:
            type: string
            description: Log
            example: "Deleted"
paths:
...
  /photo:
    # uploadPhoto
    post:
      tags: ["User", "Photo"]
      operationId: uploadPhoto
      summary: "Upload a photo"
      description: Upload a photo in your account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/photo"
        required: true
      responses:
        '201':
          description: A photo has been uploaded succesfully
          content:
            application/json:
              schema:
                type: object
                description: message object
                properties:
                  message:
                    type: string
                    description: Log
                    example: "Photo uploaded"
        '400':
          $ref : "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /photo/{photo_id}/likes:
    # likePhoto
    parameters:
      - $ref: "#/components/parameters/photo_id"
    post:
      summary: Put like to a photo
      description: Put like to a photo
      tags: ["Photo", "User"]
      operationId: likePhoto
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: username
              properties:
                username:
                  $ref: "#/components/schemas/user/properties/name"
        required: true
      responses:
        '201':
          description: Successfully added in following collection
          content:
            application/json:
              schema:
                type: object
                description: message object
                properties:
                  message:  
                    type: string
                    description: Log
                    example: "Success"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /photo/{photo_id}/likes/{uid}:
    # unlikePhoto
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - $ref: "#/components/parameters/uid"
    delete:
      summary: Unlike a photo
      description: Remove your like to one of the social network photos
      tags: ["Photo", "User"]
      operationId: unlikePhoto
      responses:
        '200':
          $ref: "#/components/responses/DeleteSuccess"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /photo/{photo_id}/comment:
    #commentPhoto
    parameters:
      - $ref: "#/components/parameters/photo_id"
    post:
      summary: Add comment to a photo
      description: Add your commento to a photo
      operationId: commentPhoto
      tags: ["Photo", "User"]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/comment"
        required: true
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                description: response object
                properties:
                  message:
                    type: string
                    description: Log
                    example: "Success"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /photo/{photo_id}/comment/{comment_id}:
    # uncommentPhoto
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - $ref: "#/components/parameters/comment_id"
    delete:
      summary: Delete a comment
      description: Delete a comment from a photo
      operationId: uncommentPhoto
      tags: ["User", "Photo"]
      responses:
        '200':
          $ref: "#/components/responses/DeleteSuccess"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 
  /photo/{photo_id}:
    # deletePhoto
    parameters:
        - $ref: "#/components/parameters/photo_id"
    delete:
      summary: Delete photo
      description: Delete a photo from the site
      tags: ["User", "Photo"]
      operationId: deletePhoto
      responses:
        '200':
          $ref: "#/components/responses/DeleteSuccess"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerDown"
      security:
        - bearerAuth: [] 